httpd config: Add X-Forwarded-For to our LogFormat
For debugging and security response, we need to be able to know the
original requestor's IP, not just the Traefik IP handling the request.
This adds the content of the X-Forwarded-For header to the combined
LogFormat, which is what we pipe out.

If there is no content, Apache will log a "-"; there is no impact to
images that directly face the Internet without being behind a proxy.